# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for Nama Pemilik Kontrakan application",
        "contact": {
            "email": "zikriramdani.developer@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/floating-whatsapp": {
            "get": {
                "tags": [
                    "Floating Whatsapp"
                ],
                "summary": "Get list of floating whatsapp",
                "description": "Returns list of floating whatsapp",
                "operationId": "4d58af6495974005c946350af2b68b3d",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FloatingWhatsapp"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/hubungi-kami": {
            "get": {
                "tags": [
                    "Hubungi Kami"
                ],
                "summary": "Get list of Hubungi Kami",
                "description": "Returns list of Hubungi Kami",
                "operationId": "4dbd8953c61a634ab8945eb0c43ff9b0",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HubungiKami"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/kontrakan": {
            "get": {
                "tags": [
                    "Kontrakan"
                ],
                "summary": "Get list of kontrakan",
                "description": "Returns list of kontrakan",
                "operationId": "e8f58de4b9067e28f0e631364ae871bd",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Kontrakan"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/logos": {
            "get": {
                "tags": [
                    "Logo"
                ],
                "summary": "Get list of logo",
                "description": "Returns list of logo",
                "operationId": "6cd2ad22184fa83758fb5da5893f0917",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Logo"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/payment": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "Get list of payment",
                "description": "Returns list of payment",
                "operationId": "d44537a300b91d45da007c299822a5ff",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Payment"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/pengaduan": {
            "get": {
                "tags": [
                    "Pengaduan"
                ],
                "summary": "Create data pengaduan",
                "description": "Returns list of pengaduan",
                "operationId": "3045cf41593b9a713b4fc590b633b108",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Pengaduan"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "FloatingWhatsapp": {
                "title": "Floating Whatsapp",
                "required": [
                    "id",
                    "avatar",
                    "phone_number",
                    "account_name",
                    "chat_message",
                    "status_message"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the floating WhatsApp",
                        "type": "integer"
                    },
                    "avatar": {
                        "description": "Avatar image URL for the floating WhatsApp",
                        "type": "string",
                        "format": "binary"
                    },
                    "phone_number": {
                        "description": "Phone number for the WhatsApp account",
                        "type": "string"
                    },
                    "account_name": {
                        "description": "Name of the WhatsApp account",
                        "type": "string"
                    },
                    "chat_message": {
                        "description": "Default chat message to be sent",
                        "type": "string"
                    },
                    "status_message": {
                        "description": "Status message for the WhatsApp account",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "HubungiKami": {
                "title": "Hubungi Kami",
                "required": [
                    "id",
                    "title"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the Hubungi Kami",
                        "type": "integer"
                    },
                    "alamat": {
                        "description": "Address",
                        "type": "string"
                    },
                    "jam_kerja": {
                        "description": "Working hours",
                        "type": "string"
                    },
                    "no_telp": {
                        "description": "Telephone number",
                        "type": "string"
                    },
                    "no_wa": {
                        "description": "WhatsApp number",
                        "type": "string"
                    },
                    "link_no_wa": {
                        "description": "WhatsApp link",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email address",
                        "type": "string"
                    },
                    "embed_google_map": {
                        "description": "Google Map embed link",
                        "type": "string"
                    },
                    "link_konfirmasi_wa": {
                        "description": "WhatsApp confirmation link",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Kontrakan": {
                "title": "Kontrakan",
                "required": [
                    "id",
                    "nama",
                    "slug",
                    "deskripsi",
                    "harga_sewa",
                    "status"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the Kontrakan",
                        "type": "integer"
                    },
                    "image": {
                        "description": "Array of image URLs or paths",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "alt": {
                        "description": "Alternative text for images",
                        "type": "string"
                    },
                    "nama": {
                        "description": "Name of the kontrakan",
                        "type": "string"
                    },
                    "slug": {
                        "description": "URL-friendly slug for the kontrakan",
                        "type": "string"
                    },
                    "deskripsi": {
                        "description": "Description of the kontrakan",
                        "type": "string"
                    },
                    "keterangan": {
                        "description": "Additional notes or remarks",
                        "type": "string"
                    },
                    "harga_sewa": {
                        "description": "Array of rental prices based on duration",
                        "type": "array",
                        "items": {
                            "properties": {
                                "durasi": {
                                    "description": "Duration in months",
                                    "type": "integer"
                                },
                                "harga": {
                                    "description": "Price in IDR",
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "status": {
                        "description": "Status of the kontrakan",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Logo": {
                "title": "Logo",
                "required": [
                    "id",
                    "image",
                    "alt"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the logo",
                        "type": "integer"
                    },
                    "image": {
                        "description": "URL of the logo image",
                        "type": "string"
                    },
                    "alt": {
                        "description": "Alt text for the logo image",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Payment": {
                "title": "Payment",
                "required": [
                    "image",
                    "alt",
                    "no_rek",
                    "nama_rek",
                    "nama_bank"
                ],
                "properties": {
                    "image": {
                        "description": "URL of the payment image",
                        "type": "string"
                    },
                    "alt": {
                        "description": "Alt text for the payment image",
                        "type": "string"
                    },
                    "no_rek": {
                        "description": "Bank account number",
                        "type": "string"
                    },
                    "nama_rek": {
                        "description": "Account holder's name",
                        "type": "string"
                    },
                    "nama_bank": {
                        "description": "Name of the bank",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Pengaduan": {
                "title": "Pengaduan",
                "required": [
                    "id",
                    "nama",
                    "no_telp",
                    "id_kontrakan",
                    "status"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the Pengaduan",
                        "type": "integer"
                    },
                    "nama": {
                        "description": "Name of the person making the complaint",
                        "type": "string"
                    },
                    "no_telp": {
                        "description": "Phone number of the person making the complaint",
                        "type": "string"
                    },
                    "alamat": {
                        "description": "Address",
                        "type": "string"
                    },
                    "catatan": {
                        "description": "Additional notes or remarks",
                        "type": "string"
                    },
                    "status": {
                        "description": "Status of the complaint",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Floating Whatsapp",
            "description": "Floating Whatsapp"
        },
        {
            "name": "Hubungi Kami",
            "description": "Hubungi Kami"
        },
        {
            "name": "Kontrakan",
            "description": "Kontrakan"
        },
        {
            "name": "Logo",
            "description": "Logo"
        },
        {
            "name": "Payment",
            "description": "Payment"
        },
        {
            "name": "Pengaduan",
            "description": "Pengaduan"
        }
    ]
}